+++ src/_/parse.ts
@@ -1,10 +1,19 @@
 /* eslint max-lines: "off", no-shadow: [ "error", { "allow": [ "focus" ] } ] */
 import type { Mode, Placeholder, Transition } from "./types";
-import { logError, regExpFinderFactory, trimRegExpFactory } from "./utils";
+import { logWarning, trimRegExpFactory } from "./utils";
 import { config } from "./install";
 import { rValidMode, rValidPlaceholder, rValidRatio } from "./validation";
 
 const rImage = /^(image:)?\/?/;
+
+const regExpFinderFactory = < T = string >( regExp: RegExp, filter: ( ( value: T ) => T ) = undefined ) =>
+    ( value: T | string ): T => {
+        let found;
+        if ( value ) {
+            ( value as string ).replace( regExp, ( _, v ) => ( found = v ) );
+        }
+        return filter ? filter( found ) : found;
+    };
 
 const isPositiveNumber = ( value: number ) => !isNaN( value ) && ( value > 0 );
 const trimOrUndefined = regExpFinderFactory( trimRegExpFactory( `.+?` ) );
@@ -19,11 +28,12 @@ export const parseFocus = trimOrUndefined;
 
 export const parseMode = regExpFinderFactory< Mode >( rValidMode );
 
-export const parsePlaceholder = regExpFinderFactory< Placeholder >(
-    rValidPlaceholder,
-    // eslint-disable-next-line no-nested-ternary
-    ( value: Placeholder ) => ( value ? ( ( value === `none` ) ? undefined : value ) : `preview` )
-);
+export const parsePlaceholder = ( placeholder: Placeholder, src:string ) : Placeholder => {
+    if ( ( config.mode === `offline` ) || !trimOrUndefined( src ) || ( placeholder === `none` ) ) {
+        return undefined;
+    }
+    return rValidPlaceholder.test( placeholder ) ? placeholder : `preview`;
+};
 
 export const parsePosition = trimOrUndefined;
 
@@ -56,10 +66,11 @@ export const parseSrc = ( value: string ): string => {
     // eslint-disable-next-line no-param-reassign
     value = trimOrUndefined( value );
     if ( !value ) {
-        logError( `src is mandatory` );
-        return undefined;
+        logWarning( `src is not provided` );
     }
-    return value.replace( rImage, `image:${ config.path }` );
+    // eslint-disable-next-line no-nested-ternary
+    return config.mode === `offline` ?
+        `` : ( value ? value.replace( rImage, `image:${ config.path }` ) : `placeholder:red` );
 };
 
 export const parseTransition = ( value: boolean | string ): Transition[] => {