+++ src/_/types.ts
@@ -1,9 +1,9 @@
 import type { PlaceholderData } from "./placeholder";
 
 export type Mode = `contain` | `cover`;
-
 export type Placeholder = `maincolor` | `meancolor` | `none` | `preview`;
 export type Transition = `fade` | `zoom` | `none`;
+export type TwicMode = `debug` | `offline` | `production`;
 
 export interface Attributes {
     alt?: string,
@@ -24,8 +24,10 @@ export interface Attributes {
 export interface Options {
     anticipation?: number,
     class?: string,
+    debug?: boolean,
     domain: string,
     maxDPR?: number,
+    mode?:TwicMode,
     path?: string,
     step?: number,
 }