+++ src/_/validation.ts
@@ -1,4 +1,4 @@
-import type { Mode, Placeholder } from "./types";
+import type { Mode, Placeholder, TwicMode } from "./types";
 import { trimRegExpFactory } from "./utils";
 
 export const validModes: Array< Mode > = [ `contain`, `cover` ];
@@ -8,3 +8,6 @@ export const validPlaceholders: Array< Placeholder > = [ `maincolor`, `meancolor
 export const rValidPlaceholder = trimRegExpFactory( validPlaceholders );
 
 export const rValidRatio = trimRegExpFactory( `(\\d+(?:\\.\\d+)?)(?:\\s*\\/\\s*(\\d+(?:\\.\\d+)?))?` );
+
+export const validTwicMode: Array< TwicMode > = [ `debug`, `offline`, `production` ];
+export const rValidTwicMode = trimRegExpFactory( validTwicMode );